Now using floats instead of doubles for audio samples.

+++ ProceduralMusic/Echo.cs
@@ -6,25 +6,25 @@ namespace ProceduralMusic
     {
         private Source _src;
         private double _delay;
-        private double _decay;
+        private float _decay;
 
-        public Echo(Source src, double delay, double decay)
+        public Echo(Source src, double delay, float decay)
         {
             _src = src;
             _delay = delay;
             _decay = decay;
         }
 
-        public override double Sample(double t)
+        public override float Sample(double t)
         {
             var reps = Math.Floor(t / _delay);
-            var vol = 1.0;
-            var sum = 0.0;
+            var vol = 1.0f;
+            var sum = 0.0f;
 
             for (int i = 0; i <= reps; ++i) {
                 sum += _src.Sample(t - i * _delay) * vol;
                 vol *= _decay;
-                if (vol < 0.01) break;
+                if (vol < 0.01f) break;
             }
 
             return sum;
